# Conflicts:
#	.idea/compiler.xml
#	.idea/misc.xml
#	.idea/modules.xml
#	app/src/main/java/com/thathustudio/spage/fragments/SubjectsFragment.java

+++ app/src/main/java/com/thathustudio/spage/adapter/PostAdapter.java
@@ -276,7 +276,7 @@ public class PostAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder> {
 
     OnLikeClickListener onLikeClickListener;
 
-    String formatTime(int time)
+    String formatTime(long time)
     {
         return DateTimeUtil.timestampToString(time);
     }
